Fixed type error for ssim-threshold arg

+++ stylize.py
@@ -36,7 +36,7 @@ parser.add_argument('--style-size', type=int, default=512,
                     help='New (minimum) size for the style image, keeping the original size if set to 0')
 parser.add_argument('--crop', type=int, default=0,
                     help='If set to anything else than 0, center crop of this size will be applied to the content image after resizing in order to create a squared image (default: 0)')
-parser.add_argument('--ssim-threshold', type=int, default=0.4, help="SSIM threshold- images below this threshold are regenerated (Default: 0.4)")
+parser.add_argument('--ssim-threshold', type=float, default=0.4, help="SSIM threshold- images below this threshold are regenerated (Default: 0.4)")
 parser.add_argument('--n_retries', type=int, default=20, help="Number of times to re-attempt stylization before taking the best image from the past N stylizations (Default: 10)")
 # random.seed(131213)
